전체 페이지에서 Appbar 구현. 권한이 있어야 페이지 접근 기능은 아직 구현하지 못함
