+++ front-end/src/js/actions/Authentication.js
@@ -9,21 +9,24 @@ import {
 	AUTH_GET_STATUS_FAILURE
 } from './ActionTypes';
 
+/* 액션생성자 함수와 thunk를 정의하는 파일*/
 
-/*로그인을 다룸*/
+/*-------------------------------로그인 & 로그아웃을 다루는 thunk 함수------------------------------*/
 export function loginRequest(email, password) {
 	return (dispatch) => {
-		/*로그인이 시작됨*/
+		/*로그인이 시작된다는 액션 객체를 리듀서로 보낸다.*/
 		dispatch(login());
 		
 		return axios.get("http://localhost:8090/admin/" + email)
 				.then((res) => {
 					if(res.data.email === email && res.data.password === password) {
+						/*로그인이 성공했다는 액션 객체를 리듀서로 보낸다.*/
 						dispatch(loginSuccess(email));
 						axios.post("http://localhost:8090/logInfo", {
 							email: email
 						})
 					} else {
+						/*로그인이 실패했다는 액션객체를 리듀서로 보낸다.*/
 						dispatch(loginFailure());
 					}
 				}).catch((error) => {
@@ -40,7 +43,10 @@ export function logoutRequest(email) {
 		});
 	};
 }
+/*---------------------------------------------------------------------------------*/
 
+
+/* ---------------------------------------로그인과 관련된 액션 생성자 함수  ------------*/
 export function logout() {
 	return {
 		type: AUTH_LOGOUT
@@ -65,8 +71,9 @@ export function loginFailure() {
 		type: AUTH_LOGIN_FAILURE
 	};
 }
+/*-------------------------------------------------------------------------*/
 
-/*로그인 정보를 가져옴 */
+/*----------------------로그인 정보 확인과 관련된 액션생성자 함수 -----------------------*/
 export function getStatusRequest() {
 	return (dispatch) => {
 		//로그인 정보 관련 API가 실행된다는 것을 알려줌.
@@ -74,6 +81,7 @@ export function getStatusRequest() {
 		
 		return axios.get("http://localhost:8090/logInfo")
 		.then((res) => {
+			//res가 존재한다고 리턴받으면 로그인 정보 조회에 성공했다는  액션객체를 dispatch 함수를 통해 리듀서로 보냄.
 			if(res.status === 200) {
 				dispatch(getStatusSuccess(res.data[0].email));
 			} else {
