+++ front-end/src/js/container/Login.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import Auth from '../component/Auth';
-import { connect } from 'react-redux';
+import { connect } from 'react-redux'; // Login 컨테이너와 리덕스를 연결해줌.
 import { loginRequest } from '../actions/Authentication'; 
 import '../css/custom.css';
 
@@ -15,8 +15,9 @@ class Login extends Component {
 						email: email
 					};
 					
+					//로그인이 성공하면 브라우저 쿠키에 로그인 정보를 저장함.
 					document.cookie = 'key=' + btoa(JSON.stringify(loginData));
-					this.props.history.push('/');
+					this.props.history.push('/users');
 					return true;
 				} else {
 					return false;	
@@ -35,6 +36,7 @@ class Login extends Component {
 	}
 }
 
+/*----------------리덕스의 state와 thunk 함수를 Login 컴포넌트로 들어온 props처럼 사용할 수 있게 함.-----------------*/
 const mapStateToProps = (state) => {
 	return {
 		status: state.Authentication.login.status
@@ -48,5 +50,6 @@ const mapDispatchToProps = (dispatch) => {
 		}
 	};
 };
+/*---------------------------------------------------------------------------------------------*/
 
 export default connect(mapStateToProps, mapDispatchToProps)(Login);
